Add commit search
This adds commits to the searched types. We filter out commits that have
corresponding PRs (since those PRs are likely to mention `msgspec`
themselves anyway). Unfortunately commit search isn't exposed through
graphql yet, so we have to make use of both the REST and graphql APIs to
get the info we want.